requirements.txt
python-binance==1.0.19
pydantic==2.5.0
click==8.1.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests==2.31.0
cryptography>=41.0.0
